git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/LUCENE-2878@1412849 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/search/MatchAllDocsQuery.java
@@ -69,7 +69,7 @@ public class MatchAllDocsQuery extends Query {
     }
 
     @Override
-    public float freq() {
+    public int freq() {
       return 1;
     }
