+++ lucene/core/src/java/org/apache/lucene/search/ReqOptSumScorer.java
@@ -94,7 +94,7 @@ class ReqOptSumScorer extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     // we might have deferred advance()
     score();
     return (optScorer != null && optScorer.docID() == reqScorer.docID()) ? 2 : 1;
