+++ lucene/join/src/java/org/apache/lucene/search/join/ToChildBlockJoinQuery.java
@@ -170,7 +170,7 @@ public class ToChildBlockJoinQuery extends Query {
     private final Bits acceptDocs;
 
     private float parentScore;
-    private float parentFreq = 1;
+    private int parentFreq = 1;
 
     private int childDoc = -1;
     private int parentDoc;
@@ -259,7 +259,7 @@ public class ToChildBlockJoinQuery extends Query {
     }
 
     @Override
-    public float freq() throws IOException {
+    public int freq() throws IOException {
       return parentFreq;
     }
