+++ lucene/queries/src/java/org/apache/lucene/queries/function/ValueSourceScorer.java
@@ -94,7 +94,7 @@ public class ValueSourceScorer extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     return 1;
   }
 }
